Attach transitionend event to specific FlipCountdown card
There was a problem in the code where it was traversing the DOM tree
to find the first counter in that tree and attach the event to it.
The issue that it caused was that if there was more than one counter
present at the same time only one of them (the first one) would
get the event attached. To fix it I simply used a ref to target
a specific element in the component and not to depend on the DOM API's
querySelector